+++ Changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed (変更点)    
 - Tauriのバージョンを2.9.3から2.10.2にアップデート  
+- Macのフルスクリーンモードの挙動を変更（Simple Fullscreen）  
 
 ### Fixed (修正点)  
 - スポットライトモードの不具合を修正